Store ranges and symbols separately for each document
This allows to keep cached ranges and symbols when editing several org
files at once, switching between them will not overwrite the other
file's cached ranges and symbols.

`TextDocument.version` is used to detect content changes. This is less
accurate (because undo and redo actions increment the version number
too), but does not require to have a copy of the whole document.